requirements.txt
httpx>=0.27
click>=8.1
pyyaml>=6.0
playwright>=1.40
rich>=13.0
